int to long

+++ src/main/java/com/chitfund/controller/IChitFundController.java
@@ -41,7 +41,7 @@ public interface IChitFundController {
 	List<Customer> getAllCustomers();
 	
 	@GetMapping("/calculateChit")
-	CalculateDto calculateChit(@RequestParam("id")int i);
+	CalculateDto calculateChit(@RequestParam("id")long i);
 	
 	@PostMapping("/updateChitFundOtherSource")
 	public ExistingChitFund updateChitFund(@RequestBody ExistingChitFund existingChitFund);
